Update and rename compress.R to reduce.R
I changed the term "data compression" to "data reduction". We've used the term "reduction" in the manuscript, and I wanted to be consistent here. When people in applied math refer to data compression, they're usually referring to the process of zipping a file (lossless compression) or storing each data point with fewer digits or lower precision to reduce memory (lossy compression). We're doing something a little bit different because we actually remove data and generate an entirely new dataset using kriging.